The script takes care of the NinitePro application with the scheduler help.
The script can update the local copy of the application ninitepro.exe.
The script will select the fastest cache location (if many will be
delivered) or will use Ninite site if caches are unavailable.
The script writes its own logs and installed applications.
With support for multiple parameters, the script works best with
Scheduler.
Scipt will be independent of the network infrastructure when you first
create chedler task and copying ninitepro.exe application to local disk
(for example with GPO).
Task schedule is best to configure this by this website:
https://brashear.me/blog/2015/01/23/deploy-application-updates-with-ninite-with-a-cache-on-your-lan/